[maybe] package-build-expand-files-spec: Support renaming files explicitly
At the same time remove support for implicitly renaming "*.el.in"
files to "*.el".  It is better if package maintainers are aware
that they are doing something out of the ordinary by having to
deal with it explicitly.

Currently this affect a whooping number of two packages on Melpa.

IMO we should not implicitly rename files.  On the other if we make
it possible to make arbitrary renames, then package authors might
actually start using that for other renames, which would be very
unfortunate.  No renames should ever be necessary and my preference
would be to get the two remaining packages that rely on the .el.in
to .el renaming to stop doing that, and then drop support for this
completely.