[WIP] Include all maintainers in archive-contents
Set an additional property `:maintainers' to the list of maintainers.
Continue to set `:maintainer' to the first of the maintainers.  We
cannot use a list of maintainers as the value of `:maintainer' because
that would break existing versions of Package.  We have to continue to
set `:maintainer' because otherwise older versions of Package would
stop displaying at least one of the maintainers.

Elpa-Admin, the tool used to maintain [Non]GNU ELPA, supports multiple
maintainers for internal purposes since [1: 384e8059a116].  However it
does not yet make that information available to Package.

Package's `package-buffer-info' supports multiple maintainers since
[2: 4e6f98cd505].  However `describe-package' still does not support
displaying multiple maintainers.

I have opened [3] in the hope that these limitations can be overcome.

TODO Some these details have to be worked out here, once [3] or
something similar has been merged.

1: 2021-08-13 384e8059a11625ecbfad22bb295ba23bd15a2e41

2: 2021-08-09 4e6f98cd505ed56e7928fcdb5ab88c09d0e74d89
   * lisp/emacs-lisp/package.el (package-buffer-info): Use lm-maintainers

3: https://debbugs.gnu.org/cgi/bugreport.cgi?bug=62524